# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /user/createUser:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 新增用户
      tags:
      - 用户模块
  /user/deleteUser:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/findUserByNameAndPassword:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 按用户名和密码查找
      tags:
      - 用户模块
  /user/getUserList:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 所有用户
      tags:
      - 用户模块
  /user/updateUser:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 用户 ID
        in: formData
        name: id
        type: string
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      - description: 邮箱
        in: formData
        name: email
        type: string
      - description: 电话
        in: formData
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户模块
swagger: "2.0"